Impedindo que uma categoria, já associada a produto, seja excluída.

+++ app/Services/ProductCategory/DeleteProductCategoryService.php
@@ -3,6 +3,7 @@
 namespace App\Services\ProductCategory;
 
 use App\Contracts\ProductCategoryRepositoryInterface;
+use App\Exceptions\ProductCategoryWithProductsException;
 
 class DeleteProductCategoryService
 {
@@ -18,6 +19,14 @@ class DeleteProductCategoryService
 
     public function execute(int $id)
     {
+        $product_category = $this->product_categories_repository->getById($id);
+
+        $product_category->loadCount('products');
+
+        if ($product_category->products_count > 0) {
+            throw new ProductCategoryWithProductsException('Não é possível excluir uma categoria de produto que já esteja associada a Produtos', 403);
+        }
+
         return $this->product_categories_repository->destroy($id);
     }
 }
